Renamed BloscWrapper to JBlosc and included init() inside constructor.
